+++ TranslateCS2.Mod/Containers/AModRuntimeContainer.cs
@@ -18,11 +18,9 @@ public abstract class AModRuntimeContainer : IModRuntimeContainer {
 
     public virtual LocalizationManager? LocManager { get; }
     public virtual InterfaceSettings? IntSetting { get; }
-    public virtual bool DoubleCheck { get; }
 
 
-    protected AModRuntimeContainer(bool doubleCheck, ILog? logger, Paths paths) {
-        this.DoubleCheck = doubleCheck;
+    protected AModRuntimeContainer(ILog? logger, Paths paths) {
         this.Logger = logger;
         this.DropDownItems = new DropDownItems();
         this.Paths = paths;
@@ -31,8 +29,7 @@ public abstract class AModRuntimeContainer : IModRuntimeContainer {
         this.ErrorMessages = new ErrorMessages(this);
         this.Languages = new MyLanguages(this);
     }
-    protected AModRuntimeContainer(bool doubleCheck, Paths paths) {
-        this.DoubleCheck = doubleCheck;
+    protected AModRuntimeContainer(Paths paths) {
         this.DropDownItems = new DropDownItems();
         this.Paths = paths;
         // the following need the Paths to be initialized!