+++ TranslateCS2.ZModDevHelper/HelpMeWithModSettingsFlavors.cs
@@ -15,31 +15,12 @@ public class HelpMeWithModSettingsFlavors {
         IEnumerable<SystemLanguage> systemLanguages = Enum.GetValues(typeof(SystemLanguage)).OfType<SystemLanguage>();
         StringBuilder builder = new StringBuilder();
         foreach (SystemLanguage systemLanguage in systemLanguages) {
-            if (systemLanguage is SystemLanguage.Unknown or SystemLanguage.Chinese) {
+            if (systemLanguage is SystemLanguage.Chinese) {
                 continue;
             }
-            builder.AppendLine($"private string _Flavor{systemLanguage} = InitFlavor(SystemLanguage.{systemLanguage});");
-            builder.AppendLine($"[Include]");
-            builder.AppendLine($"[SettingsUIDropdown(typeof(ModSettings), nameof(GetFlavors{systemLanguage}))]");
-            builder.AppendLine($"[SettingsUISection(Section, FlavorGroup)]");
-            builder.AppendLine($"[SettingsUIHideByCondition(typeof(ModSettings), nameof(IsFlavor{systemLanguage}Hidden))]");
-            builder.AppendLine($"[SettingsUIDisableByCondition(typeof(ModSettings), nameof(IsFlavor{systemLanguage}Disabled))]");
-            builder.AppendLine($"public string Flavor{systemLanguage} {{");
-            builder.AppendLine($"    get => this._Flavor{systemLanguage};");
-            builder.AppendLine($"    set => this._Flavor{systemLanguage} = this.GetValueToSet(SystemLanguage.{systemLanguage}, value);");
-            builder.AppendLine($"}}");
             builder.AppendLine($"public DropdownItem<string>[] GetFlavors{systemLanguage}() {{");
             builder.AppendLine($"    return GetFlavors(SystemLanguage.{systemLanguage});");
             builder.AppendLine($"}}");
-            builder.AppendLine($"public bool IsFlavor{systemLanguage}Hidden() {{");
-            builder.AppendLine($"    return IsHidden(SystemLanguage.{systemLanguage});");
-            builder.AppendLine($"}}");
-            builder.AppendLine($"public bool IsFlavor{systemLanguage}Disabled() {{");
-            builder.AppendLine($"    return IsDisabled(SystemLanguage.{systemLanguage});");
-            builder.AppendLine($"}}");
-            builder.AppendLine();
-            builder.AppendLine();
-            builder.AppendLine();
         }
         string text = builder.ToString()
             //.ReplaceLineEndings("\n")
@@ -51,10 +32,16 @@ public class HelpMeWithModSettingsFlavors {
         IEnumerable<SystemLanguage> systemLanguages = Enum.GetValues(typeof(SystemLanguage)).OfType<SystemLanguage>();
         StringBuilder builder = new StringBuilder();
         foreach (SystemLanguage systemLanguage in systemLanguages) {
-            if (systemLanguage is SystemLanguage.Unknown or SystemLanguage.Chinese) {
+            if (systemLanguage is SystemLanguage.Chinese) {
                 continue;
             }
-            builder.AppendLine($"this.AddToDictionary(this.modSettings.GetOptionLabelLocaleID(nameof(ModSettings.Flavor{systemLanguage})), this.GetLabel(SystemLanguage.{systemLanguage}), true);");
+            if (systemLanguage is SystemLanguage.Unknown) {
+                // take care of Unkown!
+                // it should not follow this logic!
+                // TODO: !!!
+                Assert.False(false);
+            }
+            builder.AppendLine($"this.AddToDictionary(this.modSettings.GetOptionLabelLocaleID($\"{{ModSettings.Flavor}}{{SystemLanguage.{systemLanguage}}}\"), this.GetLabel(SystemLanguage.{systemLanguage}), true);");
         }
         string text = builder.ToString()
             //.ReplaceLineEndings("\n")
@@ -66,10 +53,16 @@ public class HelpMeWithModSettingsFlavors {
         IEnumerable<SystemLanguage> systemLanguages = Enum.GetValues(typeof(SystemLanguage)).OfType<SystemLanguage>();
         StringBuilder builder = new StringBuilder();
         foreach (SystemLanguage systemLanguage in systemLanguages) {
-            if (systemLanguage is SystemLanguage.Unknown or SystemLanguage.Chinese) {
+            if (systemLanguage is SystemLanguage.Chinese) {
                 continue;
             }
-            builder.AppendLine($"this.AddToDictionary(this.modSettings.GetOptionDescLocaleID(nameof(ModSettings.Flavor{systemLanguage})), this.GetDescription(SystemLanguage.{systemLanguage}), true);");
+            if (systemLanguage is SystemLanguage.Unknown) {
+                // take care of Unkown!
+                // it should not follow this logic!
+                // TODO: !!!
+                Assert.False(false);
+            }
+            builder.AppendLine($"this.AddToDictionary(this.modSettings.GetOptionDescLocaleID($\"{{ModSettings.Flavor}}{{SystemLanguage.{systemLanguage}}}\"), this.GetDescription(SystemLanguage.{systemLanguage}), true);");
         }
         string text = builder.ToString()
             //.ReplaceLineEndings("\n")