+++ TranslateCS2.ZModDevHelper/ModTestRuntimeContainer.cs
@@ -10,32 +10,40 @@ internal class ModTestRuntimeContainer : AModRuntimeContainer {
     private static string TranslateCS2ModPropsPath { get; } = "..\\..\\..\\TranslateCS2.Inf\\Properties\\TranslateCS2.Mod.props";
     private static string EnvVariableNameManagedPath { get; } = "CSII_MANAGEDPATH";
     private static string EnvVariableNameToolPath { get; } = "CSII_TOOLPATH";
+    private static string EnvVariableNameUserDataPath { get; } = "CSII_USERDATAPATH";
     private static string GameDll { get; } = "Game.dll";
     private static string ModProps { get; } = "Mod.props";
 
 
-    public ModTestRuntimeContainer() : base(true, new Paths(true,
-                                                                             GetStreamingDataPathFromProps())) { }
+    public ModTestRuntimeContainer() : base(new Paths(true,
+                                                            GetStreamingDataPathFromProps(),
+                                                            GetUserDataPathFromEnvironment())) { }
 
 
-    private static string GetStreamingDataPathFromProps() {
-        string managedPath = Environment.GetEnvironmentVariable(EnvVariableNameManagedPath, EnvironmentVariableTarget.User);
-        if (ExistsFileInPath(managedPath, GameDll)) {
-            return JustifyPath(managedPath);
+    private static string? GetStreamingDataPathFromProps() {
+        string? managedPath = Environment.GetEnvironmentVariable(EnvVariableNameManagedPath, EnvironmentVariableTarget.User);
+        if (managedPath != null
+            && ExistsFileInPath(managedPath, GameDll)) {
+            string managedPathJustified = JustifyPath(managedPath);
+            return Paths.NormalizeUnix(managedPathJustified);
         }
-        string toolPath = Environment.GetEnvironmentVariable(EnvVariableNameToolPath, EnvironmentVariableTarget.User);
-        if (ExistsFileInPath(toolPath, ModProps)) {
+        string? toolPath = Environment.GetEnvironmentVariable(EnvVariableNameToolPath, EnvironmentVariableTarget.User);
+        if (toolPath != null
+            && ExistsFileInPath(toolPath, ModProps)) {
             string modProps = Path.Combine(toolPath, ModProps);
-            string steamDefaultManagedPath = GetPropertyValueFromXml(modProps, ".//SteamDefaultManagedPath");
-            if (ExistsFileInPath(steamDefaultManagedPath, GameDll)) {
-                return JustifyPath(steamDefaultManagedPath);
+            string? steamDefaultManagedPath = GetPropertyValueFromXml(modProps, ".//SteamDefaultManagedPath");
+            if (steamDefaultManagedPath != null
+                && ExistsFileInPath(steamDefaultManagedPath, GameDll)) {
+                string steamDefaultManagedPathJustified = JustifyPath(steamDefaultManagedPath);
+                return Paths.NormalizeUnix(steamDefaultManagedPathJustified);
             }
         }
-        string customManagedPath = GetPropertyValueFromXml(TranslateCS2ModPropsPath, ".//CustomManagedPath");
-        return JustifyPath(customManagedPath);
+        string? customManagedPath = GetPropertyValueFromXml(TranslateCS2ModPropsPath, ".//CustomManagedPath");
+        string customManagedPathJustified = JustifyPath(customManagedPath);
+        return Paths.NormalizeUnix(customManagedPathJustified);
     }
 
-    private static string GetPropertyValueFromXml(string path, string query) {
+    private static string? GetPropertyValueFromXml(string path, string query) {
         XmlDocument xmlDoc = new XmlDocument();
         xmlDoc.Load(path);
         return xmlDoc.DocumentElement.SelectSingleNode(query).InnerText;
@@ -44,12 +52,14 @@ internal class ModTestRuntimeContainer : AModRuntimeContainer {
         string gameDll = Path.Combine(path, file);
         return File.Exists(gameDll);
     }
-    public static string JustifyPath(string path) {
+    private static string? GetUserDataPathFromEnvironment() {
+        return Environment.GetEnvironmentVariable(EnvVariableNameUserDataPath, EnvironmentVariableTarget.User);
+    }
+    private static string JustifyPath(string path) {
         return
             Path
                 .Combine(path,
                          "..",
-                         "StreamingAssets")
-                .Replace("\\", "/");
+                         "StreamingAssets");
     }
 }