+++ TranslateCS2/Assets/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 [releases](https://github.com/suluknumoh/TranslateCS2/releases)
 
+## Version 2.0.1.0
+- export merge-values
+- languages
+- monospace font for mod-readme and mod-changelog
+
 ## Version 2.0
 - sessions view
     - edit/create new session